Handle text blocks in tables
Make man2fhtml understand the tbl T{ ... T} text block syntax.
This syntax is needed to get the correct line wrapping when
processing man pages with groff.  Handle by grouping lines
until the closing bracket is found.  In this implementation,
the opening bracket should be the last item on its line and
the closing one should be on its own.